Guessgame.py
This is a game called as 'Three Cup Monte'.  We usually play this game with three cups and a ball.  We place the ball in one of the cups and will shuffle and ask others to find the ball.  Here I have made the code by taking the empty cups as empty character and ball as 'o'.